Componentize makers dashboard
